Upgrades to Django 1.10.x and Wagtail 1.8.1
Along with upgrading Django and Wagtail this change upgrades other
requirements, removes django-compressor and easy-thumbnails, which are
not needed anymore as Wagtail supports Jinja2 out of the box and does
not depend on django-compressor. Therefore easy-thumbnails (which
doesn't support recent Django btw) was replaced by Wagtail's own "image"
tags.

In addition to this contains some clean-ups and small improvements.

Upgrade is a bit nuanced in terms of migrations so steps will be
communicated separately.